requirements.txt
Django==5.0.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
psycopg2==2.9.9
django-extensions==3.2.3
python-dotenv==1.0.1
pytesseract
pymupdf
pillow
drf-yasg==1.21.7
setuptools